Fix: Support VBox 6 & 7, Screensize and Video RAM
VBoxManage v6 and v7 have different VBoxManage parameter syntax...
Ubuntu default ships with V6.1, but users may have installed V7
The script checks for the version, and provides the correct syntax
for each version

On start the screen is 1024x768, which is too small for the Calamares
installer to display everything correctly. In addition on such a small
screensize the Slide show doesnt fit in the window.
We fix this by waiting 10 seconds for the VM to start and then
resizing the display.